reorganize utils functions
