+++ dynamic-mapping-ui/src/mapping/core/facade/facade-inventory.service.ts
@@ -20,7 +20,7 @@
  */
 import { Injectable } from '@angular/core';
 import { IManagedObject, InventoryService, IResult } from '@c8y/client';
-import { ProcessingContext } from '../../processor/processor.model';
+import { ProcessingContext } from '../processor/processor.model';
 import { MockInventoryService } from '../mock/mock-inventory.service';
 
 @Injectable({ providedIn: 'root' })
