+++ dynamic-mapping-ui/src/mapping/core/facade/facade-measurement.service.ts
@@ -25,7 +25,7 @@ import {
   IResult,
   MeasurementService
 } from '@c8y/client';
-import { ProcessingContext } from '../../processor/processor.model';
+import { ProcessingContext } from '../processor/processor.model';
 import { HttpStatusCode } from '@angular/common/http';
 
 @Injectable({ providedIn: 'root' })
