+++ dynamic-mapping-ui/src/mapping/core/mqtt-client.service.ts
@@ -23,7 +23,7 @@ import { Injectable } from '@angular/core';
 import { AlertService } from '@c8y/ngx-components';
 import { FacadeIdentityService } from './facade/facade-identity.service';
 import { FacadeInventoryService } from './facade/facade-inventory.service';
-import { ProcessingContext } from '../processor/processor.model';
+import { ProcessingContext } from './processor/processor.model';
 import { FacadeAlarmService } from './facade/facade-alarm.service';
 import { FacadeEventService } from './facade/facade-event.service';
 import { FacadeMeasurementService } from './facade/facade-measurement.service';
